Fix circular dependency error
